% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependent_variables.R
\name{check_dependent_variables}
\alias{check_dependent_variables}
\title{A standardization function}
\usage{
check_dependent_variables(data)
}
\arguments{
\item{data}{is a data frame containing variables NTSFD,TSFD,NTSFM,TSFM,NTAD,TAD}
}
\value{
returns data.frame with verification variables:chk_NTSFD, chk_TSFD, chk_TSFM, chk_NTSFM, chk_NTAD, chk_TAD
}
\description{
A standardization function
}
