% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_derived_variables.R
\name{check_derived_variables}
\alias{check_derived_variables}
\title{Checks derived variables}
\usage{
check_derived_variables(data)
}
\arguments{
\item{data}{is a data.frame with derived variables}
}
\value{
data.frame
}
\description{
Checks derived variables
}
\examples{
df<-data.frame(
STUDYID=c("CDISK-01","CDISK-01","CDISK-02","CDISK-02"),
PARAM=c("Hemoglobin","Glucose","Hemoglobin","Glucose"),
PARAMN=c(1,2,1,2),
STUDYIDN=c(1,2,1,2)
)
check_derived_variables(df)
}
