% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_format.R
\name{date_time_format}
\alias{date_time_format}
\title{A standardization function}
\usage{
date_time_format(data, col)
}
\value{
mod_column with datetime in ISO 8601 format
}
\description{
converts different datetime formats (time, hms, difftime, character, date) in a column to ISO 8601: yyyy-MM-dd'T'HH:mm:ss
}
\examples{
date_time_format()
}
