% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_class.R
\name{fix_age_class_tochs}
\alias{fix_age_class_tochs}
\title{fix class for standardization.}
\usage{
fix_age_class_tochs(
  data_set,
  class_var = "age",
  value_var = c("pop_all", "pop_male", "pop_female"),
  josa_year,
  fix_var = "death"
)
}
\arguments{
\item{data_set}{data set.}

\item{class_var}{class variable.}

\item{value_var}{value variable.}

\item{josa_year}{year investigated.}

\item{fix_var}{fixed variable}
}
\value{
a vector including values in changed classes
}
\description{
fix_age_class_tochs function will change classes of population data in standardization of community health survey result.
}
