% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_format.R
\name{match_format}
\alias{match_format}
\title{A standardization function}
\usage{
match_format(data, file, coltype)
}
\arguments{
\item{data}{is a data frame}

\item{file}{is path for specifications file that informs selection of variables}

\item{coltype}{is column type specification for baseline, demog, measurement, or dosing variables}
}
\value{
a data frame with specified variables converted to required format
}
\description{
matches for specified format(numeric or character) of the variable and converts variables to required format
}
