% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_missing_values.R
\name{match_missing_values}
\alias{match_missing_values}
\title{Rowwise comparison of columns between two data.frames (an extension of match_NA function)}
\usage{
match_missing_values(data1, data2, ...)
}
\arguments{
\item{data1}{first data.frame}

\item{data2}{second data.frame}

\item{...}{specifies grouping variables used to arrange the data}
}
\value{
for each pair of corresponding source and output columns a check column with true/false values is returned within a data.frame
}
\description{
rowwise comparison of missing values between data.frames to identify accidentally occurring missing values or to match two datasets for similarity
}
\examples{
 match_values_1<-match_missing_values(input<-data.frame(col1=c(2,3,"NA","","<3.5","2015-04-06",
 "XYZ","2015-04-06T09:10",
 "Cycle 1")),
 output<-data.frame(col1=c(2,3,"NA","","<3.5","2015-04-06T",
 "XYZ dosing","2015-04-06 09:10",
 "Cycle 1")))
}
