% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_ATPT.R
\name{recode_ATPT}
\alias{recode_ATPT}
\title{A standardization function}
\usage{
recode_ATPT(data, string, value)
}
\arguments{
\item{data}{}
}
\description{
A standardization function
}
\examples{
recode_ATPT()
}
