% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_VISIT.R
\name{recode_VISIT}
\alias{recode_VISIT}
\title{Numeric recoding of VISIT variable}
\usage{
recode_VISIT(data, cycle_length)
}
\arguments{
\item{data}{is a data frame containing VISIT variable for recoding}

\item{cycle_length}{is a numeric vector}
}
\value{
VISITDY when VISIT column is present in data frame
}
\description{
Numeric recoding of VISIT variable
}
\examples{
recode_VISIT(data.frame(VISIT=c("Cycle 1 Day 1","cycle 2 day 2","week 10")),cycle_length = 21)
}
