% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardization.R
\name{std_score}
\alias{std_score}
\title{Standardized scores}
\usage{
std_score(x, remove_na = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{remove_na}{Logical value indicating whether NA values should be removed before the computation. The default value is \emph{TRUE}.}

\item{...}{Further arguments can be passed to or from other methods.}
}
\value{
A vector containing standardized scores.
}
\description{
Standardized scores
}
\examples{
m <- c(2, 3, 5, 19, 30)
std_score(m)
}
