% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_alloq.R
\name{verify_alloq}
\alias{verify_alloq}
\title{verifies for unexpected missing values in ALLOQ}
\usage{
verify_alloq(
  data,
  id_col = "STUDYID",
  AVALC_col = "AVALC",
  ALLOQ_col = "ALLOQ",
  param_col = "PARAM",
  expected_missing
)
}
\arguments{
\item{data}{data.frame to be checked}

\item{id_col}{STUDYID variable}

\item{AVALC_col}{AVALC variable}

\item{ALLOQ_col}{ALLOQ variable}

\item{param_col}{PARAM variable}

\item{expected_missing}{character vector of exclusions from AVALC that are expected to have missing values in AVALN}
}
\value{
error with missing AVALN rows and otherwise NULL
}
\description{
Verifies missing AVALC matches missing AVALN values unless AVALC is in a vector that is pre-specified
}
\examples{
df1<-data.frame(
STUDYID=paste("CDISK_00",1:3),
PARAM=c(rep("Hb",3)),
AVALC=c("BLQ","BLOQ","BLQ"),
ALLOQ=c(20.3,21.2,NA_integer_)
) 

verify_alloq(df1,expected_missing = "BLQ")
}
