% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_avaln.R
\name{verify_avaln}
\alias{verify_avaln}
\title{checks if missing values in AVALN column correspond to missing values in AVALC column.}
\usage{
verify_avaln(
  data,
  id_col = "STUDYID",
  AVALC_col = "AVALC",
  AVALN_col = "AVALN",
  param_col = "PARAM",
  expected_missing
)
}
\arguments{
\item{data}{Data.frame to be checked.}

\item{id_col}{STUDYID variable.}

\item{AVALC_col}{AVALC variable.}

\item{AVALN_col}{AVALN variable.}

\item{param_col}{PARAM variable.}

\item{expected_missing}{Character vector of AVALC data that are expected to have missing values in AVALN.}
}
\value{
Error message for AVALN rows with missing values.
}
\description{
Verifies there are no missing values in AVALN when AVALC data is present unless specified in expected_missing.
}
\examples{
df2<-data.frame(
STUDYID=paste("CDISK_00",1:7),
PARAM=c(rep("Hb",3),rep("Glucose",4)),
AVALC=c(20.3,21.2,22.5,60.5,55.7,"BLQ","Sample collected"),
AVALN=c(20.3,21.2,22.5,60.5,55.7,NA_integer_,NA_integer_)
)

verify_avaln(df2,expected_missing = c("BLQ","Sample collected"))
}
