% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_no_duplicates.R
\name{verify_no_duplicates}
\alias{verify_no_duplicates}
\title{Verify that no unexpected duplicates are in the data}
\usage{
verify_no_duplicates(
  data,
  id_col = "USUBJID",
  time_col = "TSFD",
  param_col = "PARAM",
  expected = list()
)
}
\arguments{
\item{data}{The data to check}

\item{id_col, time_col, param_col}{Character scalars for the ID, TIME, and parameter name}

\item{expected}{A list of expected duplicated values}
}
\value{
The \code{data} if there is unexpected duplication, raising an error, otherwise returns data
}
\description{
Verify that no unexpected duplicates are in the data
}
\examples{
df_wo_dup<-data.frame(
USUBJID=c(rep("CDISC01.10001",6)),
PARAM=c("pk","pk","dosing","dosing","dosing","dosing"),
TSFD=c(2542.9,25421.9,3213,2543,2456,3241),
VISIT=c("WEEK 16","WEEK 16","Week 3","Week 4","Week 5","Week 6"),
ATPT=c("","","0 hours","0 hours","0 hours","0 hours"),
ADTC=c("2020-03-27T10:01:UN","","","","","")
) 
 
verify_no_duplicates(data=df_wo_dup)
}
