% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dependent_variables.R
\name{check_dependent_variables}
\alias{check_dependent_variables}
\title{A standardization function}
\usage{
check_dependent_variables(data)
}
\arguments{
\item{data}{}
}
\description{
A standardization function
}
