% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_derived_variables.R
\name{check_derived_variables}
\alias{check_derived_variables}
\title{Ensures no duplicate parameter names exist. Compares output dataset with specifications file for number of elements in columns PARAMN and PARAM.}
\usage{
check_derived_variables(data, specs)
}
\arguments{
\item{data}{Data.frame with PARAM and PARAMN columns.}

\item{specs}{specifications file with 'Recode PARAM' sheet required.}
}
\value{
Data.frame with number of unique elements in PARAM and PARAMN.
}
\description{
Compares output dataset with specifications file for number of elements in columns PARAMN and PARAM.
}
\examples{
df<-data.frame(
PARAM=c("Hemoglobin","Glucose","Hemoglobin","Glucose"),
PARAMN=c(1,2,1,2)
)
list1<-data.frame(
PARAM=c("Hemoglobin","Glucose"),
PARAMN=c(1,2)
)
specification<-list(list1)
names(specification)<-"Recode PARAM"
check_derived_variables(df,specification)
}
