% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_merge_output.R
\name{check_merge_output}
\alias{check_merge_output}
\title{A standardization function}
\usage{
check_merge_output(data, study, ...)
}
\value{
returns duplicate rows
}
\description{
checks for dupplicate rows
}
\examples{
check_merge_output()
}
