% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_merge_output.R
\name{check_merge_output}
\alias{check_merge_output}
\title{A standardization function}
\usage{
check_merge_output(data, study, ...)
}
\arguments{
\item{data}{is a data frame}

\item{study}{is a character STUDYID pattern to check data frame for duplicate rows for each study}

\item{...}{are grouping variables}
}
\value{
throws an error and prints duplicate rows
}
\description{
checks for duplicate rows
}
