% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_units.R
\name{check_units}
\alias{check_units}
\title{A standardization function}
\usage{
check_units(data, col1, col2, exclude_var)
}
\arguments{
\item{col2}{}
}
\description{
A standardization function
}
\examples{
check_units()
}
