% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_units.R
\name{check_units}
\alias{check_units}
\title{Checks for uniformity of units within a parameter name across multiple studies.}
\usage{
check_units(data, nstudies)
}
\arguments{
\item{data}{Data.frame with STUDYID,USUBJID,PARAM,EXDOSEU,AVALU variables.}

\item{nstudies}{Specifies number of studies.}
}
\value{
Data.frame with summarized parameters and their respective units. Also returns a message for parameter names with multiple or missing units.
}
\description{
Checks for uniformity of units within a parameter name across multiple studies.
}
\examples{
vignette(package="standardization",all=TRUE)
df <- data.frame(
  STUDYID = c("CDISK-01", "CDISK-01", "CDISK-01", "CDISK-01",
              "CDISK-02", "CDISK-02","CDISK-02", "CDISK-02"),
  USUBJID = paste0("CDISC01.10000",1:8),
  PARAM = c("CDISK dosing", "CDISK dosing", "Glucose", "Glucose",
            "CDISK dosing", "CDISK dosing", "Glucose", "Glucose"),
  EXDOSEU = c("mg", "mg", NA, NA,
              "mg", "mg", NA, NA),
  EXDOSE=c(28, 28, NA, NA,
           50, 50, NA, NA),
  AVALN=c(NA, NA, 3.9 , 4,
          NA, NA, 5, NA),
  AVALU = c(NA, NA, "mmol/L", "mmol/L",
            NA, NA, "mmol/L", NA)
)
check_units(df, nstudies=2)
}
