% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_time_format.R
\name{date_time_format}
\alias{date_time_format}
\title{Title}
\usage{
date_time_format(data, col)
}
\arguments{
\item{data}{is a data frame with datetime variables}

\item{col}{is datetime variable}
}
\value{
datetime variable in ISO 8601 format with UN for missing date and time parts
}
\description{
converts datetime variable to ISO 8601
}
\examples{
date_time_format(data.frame(ADTC=c("2012-04-05T09:10")),"ADTC")
}
