% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_comparison_prior_condition.R
\name{datetime_comparison_prior_condition}
\alias{datetime_comparison_prior_condition}
\title{A standardization function}
\usage{
datetime_comparison_prior_condition(
  file_name1,
  file_name2,
  datetimecol1,
  datetimecol2,
  datetimecol3,
  colwithconditions,
  filter_val,
  condition1
)
}
\description{
compares datetime with earliest dosing datetime to extract prior condition
}
