% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_colnames.R
\name{match_colnames}
\alias{match_colnames}
\title{Matches for standard column names and their variants, selects applicable variables present in source dataset, and creates empty columns for missing variables}
\usage{
match_colnames(data, file, coltype)
}
\arguments{
\item{data}{is a data frame}

\item{file}{is path for specifications file that informs selection of variables}

\item{coltype}{is column type specification for baseline, demog, measurement, or dosing variables}
}
\value{
a data frame with specified variables selected
}
\description{
matches for standard column names and their variants present in specifications file, selects applicable variables present in source dataset, and creates empty columns for missing variables
}
