% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_colnames.R
\name{match_colnames}
\alias{match_colnames}
\title{Selects SDTM column names or their name variants and renames them to preferred column name in the output}
\usage{
match_colnames(data, file, coltype)
}
\arguments{
\item{data}{A data.frame that serves as a source dataset}

\item{file}{Specifications can be provided as list object in R or .xlsx file path, which informs selection of variables}

\item{coltype}{Column type specification: 'baseline', 'demog', 'measurement', or 'dosing' variables}
}
\value{
A data.frame with specified variables selected and renamed to preferred name
}
\description{
Selects SDTM column names or their name variants and renames them to preferred column name in the output.
}
\examples{
df<-data.frame(STUDYID=rep("S-CDSK-01",3),DOMAIN=rep("YEARS",3),Age=c(72,66,78))
list1<-data.frame(`Column Name`="AGE",
`Column Name Variants`="Age",
Labels="Age in years at baseline",
`Column Type`="Demographics",
Format="Numeric")
specification<-list(list1)
names(specification)<-"Specification-Source Data"
format_df<-match_colnames(df,file=specification,coltype="Demographics")
}
