% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_colnames.R
\name{match_colnames}
\alias{match_colnames}
\title{A standardization function}
\usage{
match_colnames(data, file, coltype)
}
\value{
data frame with selected columns
}
\description{
matches for standard column names and selects from source dataset and prints columns not present in source dataset
}
