% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_missing_values.R
\name{match_missing_values}
\alias{match_missing_values}
\title{A standardization function}
\usage{
match_missing_values(data1, data2, ...)
}
\arguments{
\item{input_var}{}
}
\description{
A standardization function
}
\examples{
match_missing_values()
}
