% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_missing_values.R
\name{match_missing_values}
\alias{match_missing_values}
\title{Rowwise comparison of missing values between two data.frames}
\usage{
match_missing_values(data1, data2, ...)
}
\arguments{
\item{data1}{first data.frame}

\item{data2}{second data.frame}

\item{...}{are grouping variables used to arrange the data}
}
\value{
for each pair of corresponding source and output columns a check column with true/false values returned within a data.frame
}
\description{
rowwise comparison of missing values between data.frames to identify accidentally occurring missing values
}
\examples{
match_missing_values(source<-data.frame(col1=c(2,3,"NA")), output<-data.frame(col1=c(2,3,"NA")))
}
