% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_missing_values.R
\name{match_missing_values}
\alias{match_missing_values}
\title{Compares output and source dataset columns and identifies data mismatches.}
\usage{
match_missing_values(input, output, ...)
}
\arguments{
\item{input}{Source dataset.}

\item{output}{Output dataset.}

\item{...}{Specifies grouping variables used to arrange and align the two datasets.}
}
\value{
Data.frame with TRUE or FALSE values to indicate matches and mismatches, respectively.
}
\description{
Identifies mismatches between corresponding source and output dataset columns.
}
\examples{
 match_values_1<-match_missing_values(input<-data.frame(col1=c(2,3,"NA","","<3.5","2015-04-06",
 "CDISK","2015-04-06T09:10",
 "Cycle 1")),
 output<-data.frame(col1=c(2,3,"NA","","<3.5","2015-04-06T",
 "CDISK dosing","2015-04-06 09:10",
 "Cycle 1")))
}
