% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_NA.R
\name{match_NA}
\alias{match_NA}
\title{Compares sum of missing values between corresponding output and source dataset columns.}
\usage{
match_NA(source, output)
}
\arguments{
\item{source}{Source dataset}

\item{output}{Output dataset}
}
\value{
Draws a comparison of total number of missing values between two datasets and returns an error message for unequal matches.
}
\description{
Compares sum of missing values between corresponding output and source dataset columns.
}
\examples{
match_NA(source<-data.frame(col1=c(2,3,"NA","")), output<-data.frame(col1=c(2,3,"NA","")))
}
