% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_match_colnames.R
\name{meta_match_colnames}
\alias{meta_match_colnames}
\title{A standardization function}
\usage{
meta_match_colnames(data, file, coltype)
}
\arguments{
\item{data}{is a data frame}

\item{file}{is path for specifications file that informs selection of variables}

\item{coltype}{is column type specification for baseline, demog, measurement, or dosing variables}
}
\value{
a data frame with specified variables selected
}
\description{
matches for variables based on metadata, selects applicable variables present in source dataset, and creates empty columns for missing variables
}
