% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_match_colnames.R
\name{meta_match_colnames}
\alias{meta_match_colnames}
\title{A standardization function}
\usage{
meta_match_colnames(data, file, coltype)
}
\arguments{
\item{coltype}{}
}
\description{
A standardization function
}
\examples{
meta_match_colnames()
}
