% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk_variables.R
\name{pk_variables}
\alias{pk_variables}
\title{A standardization function}
\usage{
pk_variables(data, col)
}
\arguments{
\item{data}{is a data frame containing AVALC variable for recoding}

\item{col}{is the equivalent for AVALC variable}
}
\value{
AVALN, ALLOQ and AULOQ
}
\description{
A standardization function
}
\examples{
pk_variables(data.frame(AVALC=c("<2.30","3.45","BLQ","BQL<2.1")),"AVALC")
}
