% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_ATPT.R
\name{recode_ATPT}
\alias{recode_ATPT}
\title{Numeric recoding of ATPT variable}
\usage{
recode_ATPT(data)
}
\arguments{
\item{data}{containing ATPT variable for recoding}
}
\value{
\code{data} with the column ATPTN
}
\description{
Numeric recoding of ATPT variable
}
\examples{
recode_ATPT(data.frame(ATPT=c("Pre-dose","2 hrs postdose","drug A (2 hours before)","34")))
}
