% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_ATPT.R
\name{recode_ATPT}
\alias{recode_ATPT}
\title{Numeric re-coding of ATPT variable}
\usage{
recode_ATPT(data)
}
\arguments{
\item{data}{A data.frame with ATPT variable.}
}
\value{
Data.frame with the addition of ATPTN variable
}
\description{
Converts planned time point data in ATPT column to numeric format
}
\examples{
recode_ATPT(data.frame(ATPT = c("Pre-dose", "2 hrs postdose", "drug A (2 hours before)", "34")))
}
