% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_blank.R
\name{replace_na_blank}
\alias{replace_na_blank}
\title{Replaces NA and blanks with replacement pattern in multiple columns.}
\usage{
replace_na_blank(data, columns, replacement)
}
\arguments{
\item{data}{is a data.frame.}

\item{columns}{specifies columns.}

\item{replacement}{Specifies columns where replacement pattern is to be applied.}
}
\value{
Data.frame where NA and blanks are replaced with specified replacement pattern.
}
\description{
Replaces NA and blanks with replacement pattern in multiple columns.
}
\examples{
replace_na_blank(data.frame(col1=c(1,"",NA_character_)),c("col1"),"U")
}
