% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_alloq.R
\name{verify_alloq}
\alias{verify_alloq}
\title{Verifies if ALLOQ values are missing when AVALC data is present}
\usage{
verify_alloq(
  data,
  id_col = "STUDYID",
  AVALC_col = "AVALC",
  ALLOQ_col = "ALLOQ",
  param_col = "PARAM",
  expected_missing
)
}
\arguments{
\item{data}{Data.frame to be checked}

\item{id_col}{STUDYID variable}

\item{AVALC_col}{AVALC variable}

\item{ALLOQ_col}{ALLOQ variable}

\item{param_col}{PARAM variable}

\item{expected_missing}{Character vector of AVALC data that are expected to have missing values in ALLOQ.}
}
\value{
Error message for data rows with missing values in ALLOQ.
}
\description{
Checks if ALLOQ values are missing when AVALC data is present.
}
\examples{
df1<-data.frame(
STUDYID=paste("CDISK_00",1:3),
PARAM=c(rep("Hb",3)),
AVALC=c("BLQ","BLOQ","BLQ"),
ALLOQ=c(20.3,21.2,NA_integer_)
)

verify_alloq(df1,expected_missing = "BLQ")
}
