% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_avaln.R
\name{verify_avaln}
\alias{verify_avaln}
\title{verifies missing AVALC matches missing AVALN values}
\usage{
verify_avaln(
  data,
  id_col = "STUDYID",
  AVALC_col = "AVALC",
  AVALN_col = "AVALN",
  param_col = "PARAM",
  expected_missing
)
}
\arguments{
\item{data}{data.frame to be checked}

\item{id_col}{STUDYID variable}

\item{AVALC_col}{AVALC variable}

\item{AVALN_col}{AVALN variable}

\item{param_col}{PARAM variable}

\item{expected_missing}{character vector of exclusions from AVALC that are expected to have missing values in AVALN}
}
\value{
error with missing AVALN rows and otherwise NULL
}
\description{
Verifies missing AVALC matches missing AVALN values unless AVALC is in a vector that is pre-specified
}
\examples{
df2<-data.frame(
STUDYID=paste("CDISK_00",1:7),
PARAM=c(rep("Hb",3),rep("Glucose",4)),
AVALC=c(20.3,21.2,22.5,60.5,55.7,"BLQ","Sample collected"),
AVALN=c(20.3,21.2,22.5,60.5,55.7,NA_integer_,NA_integer_)
)

verify_avaln(df2,expected_missing = c("BLQ","Sample collected"))  
}
